% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R, R/RenDate_20250701.R
\name{calibrate_uniform}
\alias{calibrate_uniform}
\title{calibrate_uniform}
\usage{
calibrate_uniform(
  measure.min,
  measure.max,
  calCurves,
  ids = NULL,
  positions = NULL,
  pathToCalCurves = paste0(getwd(), "/", "data"),
  timeScale = 1
)

calibrate_uniform(
  measure.min,
  measure.max,
  calCurves,
  ids = NULL,
  positions = NULL,
  pathToCalCurves = paste0(getwd(), "/", "data"),
  timeScale = 1
)
}
\arguments{
\item{measure.min}{valeur basse de mesure ou liste des mesures à calibrer}

\item{measure.max}{valeur haute de mesure ou liste des valeurs à calibrer}

\item{calCurves}{Le nom d'une courbe de calibration qui est dans le répertoire data}

\item{ids}{un index}

\item{positions}{une position en hauteur sur un graphe avec plusieurs dates}

\item{pathToCalCurves}{chemin d'acces aux fichiers de calibrations}

\item{timeScale}{pas de la grille de temps}
}
\description{
Fonction qui permet le calcul de la densité de probabilté d'une date, à partir d'une mesure avec une densité de probabilité uniforme

{Fonction qui permet le calcul de la densité de probabilté d'une date, à partir d'une mesure avec une densité de probabilité uniforme}
}
