% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/RenDate.R, R/RenDate_20250701.R
=======
% Please edit documentation in R/RenDate.R
>>>>>>> d229a2265573510b59c90251571a5d0a6449fc85
\name{CRAtoD14C}
\alias{CRAtoD14C}
\title{CRA to D14C}
\usage{
CRAtoD14C(CRA)
<<<<<<< HEAD

CRAtoD14C(CRA)
=======
>>>>>>> d229a2265573510b59c90251571a5d0a6449fc85
}
\arguments{
\item{CRA}{Âge radiocarbone}
}
\value{
Delta 14C (‰)
<<<<<<< HEAD

Delta 14C (‰)
}
\description{
Conversion d'un âge radiocarbone en Delta 14C

Conversion d'un âge radiocarbone en Delta 14C
=======
}
\description{
Conversion d'un âge radiocarbone en Delta 14C
>>>>>>> d229a2265573510b59c90251571a5d0a6449fc85
}
