% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R
\name{CRAtoF14C}
\alias{CRAtoF14C}
\title{CRA to F14C}
\usage{
CRAtoF14C(CRA)
}
\arguments{
\item{CRA}{Âge radiocarbone conventionnel}
}
\value{
Fraction moderne F14C
}
\description{
Conversion d'un âge radiocarbone (CRA) en fraction moderne (F14C)
}
