% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R, R/RenDate_20250701.R
\name{D14CtoCRA}
\alias{D14CtoCRA}
\title{D14C to CRA}
\usage{
D14CtoCRA(D14C)

D14CtoCRA(D14C)
}
\arguments{
\item{D14C}{Delta 14C}
}
\value{
CRA (창ge radiocarbone)

CRA (창ge radiocarbone)
}
\description{
Conversion de Delta 14C en 창ge radiocarbone (CRA)

Conversion de Delta 14C en 창ge radiocarbone (CRA)
}
