% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R, R/RenDate_20250701.R
\name{D14CtoF14C}
\alias{D14CtoF14C}
\title{D14C to F14C}
\usage{
D14CtoF14C(D14C)

D14CtoF14C(D14C)
}
\arguments{
\item{D14C}{Delta 14C en pour mille (‰)}
}
\value{
Fraction moderne F14C

Fraction moderne F14C
}
\description{
Conversion de Delta 14C (‰) vers F14C

Conversion de Delta 14C (‰) vers F14C
}
