% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R
\name{date.gaussian}
\alias{date.gaussian}
\title{Fonction qui permet le calcul de la densité de probabilté de date uniforme - fonction porte}
\usage{
date.gaussian(
  mean = 0,
  sd = 10,
  time.grid.min = -1000,
  time.grid.max = 2000,
  time.grid.scale = 1,
  ids = NULL,
  position = NULL
)
}
\arguments{
\item{mean}{moyenne}

\item{sd}{standard deviations}

\item{ids}{nom ou identifiant}

\item{timeGrid.min}{valeur minimale de la grille}

\item{timeGrid.max}{valeur minimale de la grille}

\item{timeScale}{pas de la grille de temps}
}
\description{
Fonction qui permet le calcul de la densité de probabilté de date uniforme - fonction porte
}
