% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R, R/RenDate_20250701.R
\name{date_gaussian}
\alias{date_gaussian}
\title{Fonction qui permet le calcul de la densité de probabilté de date gaussienne - fonction gauss}
\usage{
date_gaussian(
  mean = 0,
  sd = 10,
  time.grid.min = -1000,
  time.grid.max = 2000,
  time.grid.scale = 1,
  ids = NULL,
  position = NULL
)

date_gaussian(
  mean = 0,
  sd = 10,
  time.grid.min = -1000,
  time.grid.max = 2000,
  time.grid.scale = 1,
  ids = NULL,
  position = NULL
)
}
\arguments{
\item{mean}{moyenne}

\item{sd}{standard deviations}

\item{time.grid.min}{valeur minimale de la grille}

\item{time.grid.max}{valeur minimale de la grille}

\item{time.grid.scale}{pas de la grille de temps}

\item{ids}{nom ou identifiant}

\item{positions}{une position en hauteur sur un graphe avec plusieurs dates}

\item{timeGrid.min}{valeur minimale de la grille}

\item{timeGrid.max}{valeur minimale de la grille}

\item{timeScale}{pas de la grille de temps}
}
\description{
Fonction qui permet le calcul de la densité de probabilté de date gaussienne - fonction gauss

Fonction qui permet le calcul de la densité de probabilté de date gaussienne - fonction gauss
}
