% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R, R/RenDate_20250701.R
\name{delta14CtoCRA}
\alias{delta14CtoCRA}
\title{Delta14C to CRA}
\usage{
delta14CtoCRA(delta14C, year.calBP)

delta14CtoCRA(delta14C, year.calBP)
}
\arguments{
\item{delta14C}{Delta 14C (‰)}

\item{year.calBP}{Année en cal BP}
}
\value{
CRA (âge radiocarbone)

CRA (âge radiocarbone)
}
\description{
Conversion directe de Delta14C en CRA (âge radiocarbone)

Conversion directe de Delta14C en CRA (âge radiocarbone)
}
