% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R, R/RenDate_20250701.R
\name{delta14CtoF14C.inBC}
\alias{delta14CtoF14C.inBC}
\title{Delta14C to F14C in BC}
\usage{
delta14CtoF14C.inBC(delta14C, Year.BC)

delta14CtoF14C.inBC(delta14C, Year.BC)
}
\arguments{
\item{delta14C}{Delta 14C (‰)}

\item{Year.BC}{Année avant 1950 (BC)}
}
\value{
F14C ajusté pour une date BC

F14C ajusté pour une date BC
}
\description{
Conversion de Delta14C en F14C pour une date avant 1950 (en années BC)

Conversion de Delta14C en F14C pour une date avant 1950 (en années BC)
}
