% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R, R/RenDate_20250701.R
\name{delta14CtoF14C.inBP}
\alias{delta14CtoF14C.inBP}
\title{Delta14C to F14C in BP}
\usage{
delta14CtoF14C.inBP(delta14C, theta.calBP)

delta14CtoF14C.inBP(delta14C, theta.calBP)
}
\arguments{
\item{delta14C}{Delta 14C (‰)}

\item{theta.calBP}{Date en années cal BP}
}
\value{
F14C corrigé pour une date BP

F14C corrigé pour une date BP
}
\description{
Conversion de Delta14C en F14C pour une date en BP (cal BP)

Conversion de Delta14C en F14C pour une date en BP (cal BP)
}
