% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R, R/RenDate_20250701.R
\name{errCRAtoErrD14C}
\alias{errCRAtoErrD14C}
\title{Erreur CRA vers erreur D14C}
\usage{
errCRAtoErrD14C(CRA, errCRA)

errCRAtoErrD14C(CRA, errCRA)
}
\arguments{
\item{CRA}{Âge radiocarbone}

\item{errCRA}{Erreur sur le CRA}
}
\value{
Erreur sur Delta 14C

Erreur sur Delta 14C
}
\description{
Propagation de l'erreur du CRA vers D14C

Propagation de l'erreur du CRA vers D14C
}
