% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R
\name{errCRAtoErrF14C}
\alias{errCRAtoErrF14C}
\title{Erreur CRA vers erreur F14C}
\usage{
errCRAtoErrF14C(CRA, errCRA)
}
\arguments{
\item{CRA}{Âge radiocarbone}

\item{errCRA}{Erreur sur le CRA}
}
\value{
Erreur sur F14C
}
\description{
Propagation de l'erreur du CRA vers F14C
}
