% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R, R/RenDate_20250701.R
\name{errD14CtoErrCRA}
\alias{errD14CtoErrCRA}
\title{Erreur D14C vers erreur CRA}
\usage{
errD14CtoErrCRA(errD14C, D14C)

errD14CtoErrCRA(errD14C, D14C)
}
\arguments{
\item{errD14C}{Erreur sur Delta 14C}

\item{D14C}{Valeur de Delta 14C}
}
\value{
Erreur sur CRA

Erreur sur CRA
}
\description{
Propagation d'erreur de D14C vers CRA

Propagation d'erreur de D14C vers CRA
}
