% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R, R/RenDate_20250701.R
\name{errD14CtoErrF14C}
\alias{errD14CtoErrF14C}
\title{Erreur D14C vers erreur F14C}
\usage{
errD14CtoErrF14C(errD14C)

errD14CtoErrF14C(errD14C)
}
\arguments{
\item{errD14C}{Erreur sur Delta 14C (‰)}
}
\value{
Erreur sur F14C

Erreur sur F14C
}
\description{
Conversion de l'erreur sur Delta 14C vers erreur sur F14C

Conversion de l'erreur sur Delta 14C vers erreur sur F14C
}
