% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R
\name{errF14CtoErrCRA}
\alias{errF14CtoErrCRA}
\title{Erreur F14C vers erreur CRA}
\usage{
errF14CtoErrCRA(F14C, errF14C)
}
\arguments{
\item{F14C}{Fraction moderne}

\item{errF14C}{Erreur associée à F14C}
}
\value{
Erreur sur l'âge CRA
}
\description{
Propagation de l'erreur de F14C vers l'erreur en CRA
}
