% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R, R/RenDate_20250701.R
\name{errF14CtoErrD14C}
\alias{errF14CtoErrD14C}
\title{Erreur F14C vers erreur D14C}
\usage{
errF14CtoErrD14C(errF14C)

errF14CtoErrD14C(errF14C)
}
\arguments{
\item{errF14C}{Erreur sur F14C}
}
\value{
Erreur sur D14C (‰)

Erreur sur D14C (‰)
}
\description{
Conversion de l'erreur sur F14C vers erreur sur D14C

Conversion de l'erreur sur F14C vers erreur sur D14C
}
