% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R
\name{F14CtoD14C}
\alias{F14CtoD14C}
\title{F14C to D14C}
\usage{
F14CtoD14C(F14C)
}
\arguments{
\item{F14C}{Fraction moderne}
}
\value{
Delta 14C en pour mille (‰)
}
\description{
Conversion de F14C vers Delta 14C (‰)
}
