% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R
\name{F14CtoDelta14C}
\alias{F14CtoDelta14C}
\title{F14C to Delta14C}
\usage{
F14CtoDelta14C(F14C, Year.BC)
}
\arguments{
\item{F14C}{Fraction moderne}

\item{Year.BC}{Année en BC}
}
\value{
Delta14C (‰)
}
\description{
Conversion de F14C en Delta14C pour une année BC
}
