% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R, R/RenDate_20250701.R
\name{map}
\alias{map}
\title{map}
\usage{
map(date)

map(date)
}
\arguments{
\item{date}{densité de date de class "RenDate"}
}
\description{
{Retourne le temps correspondant au maximum d'une densité de date}

{Retourne le temps correspondant au maximum d'une densité de date}
}
\examples{
 g_dat <- date_gaussian( mean = 20, sd = 3, time.grid.scale = .5)
 'max(g_dat$`N(20;3)`))'
 g_dat <- date_gaussian( mean = 20, sd = 3, time.grid.scale = .5)
 'max(g_dat$`N(20;3)`))'
}
