% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R
\name{mesure.enveloppe}
\alias{mesure.enveloppe}
\title{Trace une roite représentant une mesure avec son enveloppe d erreur à 1 sigma et deux sigma}
\usage{
mesure.enveloppe(
  t,
  mesure,
  std,
  col.env = "gray",
  col.mesure = "darkgray",
  ...
)
}
\description{
Trace une roite représentant une mesure avec son enveloppe d erreur à 1 sigma et deux sigma
}
