% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R
\name{period_reduction}
\alias{period_reduction}
\title{Fonction utilisée pour réduire l'intervalle support d'une densité de class "RenDate"}
\usage{
period_reduction(dens, tmin = 0, tmax = 2000)
}
\arguments{
\item{dens}{une densité (datation) de class "RenDate"}

\item{tmin}{la valeur moyenne}

\item{tmax}{la valeur de l'écart type}
}
\description{
Fonction utilisée pour réduire l'intervalle support d'une densité de class "RenDate"
}
