% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R, R/RenDate_20250701.R
\name{plot_gaussian_measure}
\alias{plot_gaussian_measure}
\title{plot_gaussian_measure}
\usage{
plot_gaussian_measure(
  mean = 0,
  sd = 1,
  k = 5,
  x.min = 0,
  size = 10,
  col = "skyblue",
  border = NA,
  normalized = TRUE,
  new = F
)

plot_gaussian_measure(
  mean = 0,
  sd = 1,
  k = 5,
  x.min = 0,
  size = 10,
  col = "skyblue",
  border = NA,
  normalized = TRUE,
  new = F
)
}
\arguments{
\item{mean}{valeur moyenne de la gaussienne}

\item{sd}{ecart type}

\item{k}{facteur fixe la distance à la moyenne du dessin min = mean - k*sd, et max = mean + k*sd}

\item{x.min}{position de la base de la gaussienne en unité du graph}

\item{size}{taille horizontal de la gaussienne en unité du graph}

\item{normalized}{force la densité à avoir un maximum à 1 * size, sinon c'est la surface qui est égale à 1}
}
\description{
Trace verticalement une mesure gaussienne.

{Trace verticalement une mesure gaussienne.}
}
