% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R, R/RenDate_20250701.R
\name{plot_square_measure}
\alias{plot_square_measure}
\title{plot_square_measure}
\usage{
plot_square_measure(
  min = 50,
  max = 100,
  x.min = 0,
  size = 10,
  col = "skyblue",
  border = NA,
  normalized = TRUE
)

plot_square_measure(
  min = 50,
  max = 100,
  x.min = 0,
  size = 10,
  col = "skyblue",
  border = NA,
  normalized = TRUE
)
}
\arguments{
\item{min}{valeur basse de la mesure}

\item{max}{valeur haute de la mesure}

\item{x.min}{position de la base de la gaussienne en unité du graph}

\item{size}{taille horizontal de la gaussienne en unité du graph}

\item{normalized}{force la densité à avoir un maximum à 1 * size, sinon c'est la surface qui est égale à 1}
}
\description{
{Trace verticalement une mesure rectangulaire. porte}

{Trace verticalement une mesure rectangulaire. porte}
}
