% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R
\name{pred_value}
\alias{pred_value}
\title{Random value between Unit.B and Unit.D
$$tc  =  tb  +  Unif[0; 1] * ( td - tb )$$
  OR 
$$tc  =  Unif[tb; td]$$}
\usage{
pred_value(a, b, seed = "NA")
}
\arguments{
\item{a}{un vecteur, par exemple history plot provenant de ChronoModel}

\item{b}{un vecteur, par exemple history plot provenant de ChronoModel}
}
\value{
un vecteur
}
\description{
Random value between Unit.B and Unit.D
$$tc  =  tb  +  Unif[0; 1] * ( td - tb )$$
  OR 
$$tc  =  Unif[tb; td]$$
}
