% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R
\name{produit.RenDate}
\alias{produit.RenDate}
\title{Calcul la combinaison au sens produit de deux densités de class "RenDate"}
\usage{
produit.RenDate(date1, date2, timeScale = 1)
}
\arguments{
\item{timeScale}{permet de modifier la grille de temps}
}
\description{
Calcul la combinaison au sens produit de deux densités de class "RenDate"
}
