% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R, R/RenDate_20250701.R
\name{quantile}
\alias{quantile}
\title{quantile}
\usage{
quantile(date, prob = c(0.25, 0.5, 0.75))

quantile(date, prob = c(0.25, 0.5, 0.75))
}
\arguments{
\item{date}{densité de date de class "RenDate"}

\item{prob}{quantiles  default prob = c(0.25, 0.5, 0.75)}
}
\description{
Retourne le temps correspondant à la ou aux quantiles juste supérieures ou égale d'une densité de date.

{Retourne le temps correspondant à la ou aux quantiles juste supérieures ou égale d'une densité de date.}
}
\examples{
 'g_dat <- date_gaussian( mean = 20, sd = 3, time.grid.scale = .5)'
 'quantile(g_dat[[1]])'
 'g_dat <- date_gaussian( mean = 20, sd = 3, time.grid.scale = .5)'
 'quantile(g_dat[[1]])'
}
