% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R
\name{read.Ref}
\alias{read.Ref}
\title{Lecture d'un fichier de référence}
\usage{
read.Ref(file.Ref, encoding = "macroman")
}
\arguments{
\item{encoding}{Pour les fichiers du MacOS, il faut "macroman" -> difficle à connaitre, peut être "latin1" ou "utf8".}
}
\value{
une liste de data.frame
}
\description{
Lecture d'un fichier de référence
}
