% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R, R/RenDate_20250701.R
\name{read_ref}
\alias{read_ref}
\title{read_ref}
\usage{
read_ref(file.Ref, encoding = "macroman")

read_ref(file.Ref, encoding = "macroman")
}
\arguments{
\item{encoding}{Pour les fichiers du MacOS, il faut "macroman" -> difficle à connaitre, peut être "latin1" ou "utf8".}
}
\value{
une liste de data.frame

une liste de data.frame
}
\description{
Lecture d'un fichier de référence

Lecture d'un fichier de référence
}
\examples{
'curveI <- read.Ref("Calib/AM/GAL2002sph2014_I.ref")'
'curveI <- read.Ref("Calib/AM/GAL2002sph2014_I.ref")'
}
