% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R
\name{RenDate.hdr}
\alias{RenDate.hdr}
\title{Calcul le hdr (hpd) sur une densité de probabilité de date}
\usage{
RenDate.hdr(date, prob = 0.95)
}
\arguments{
\item{prob}{surface [0, 1]}
}
\description{
Calcul le hdr (hpd) sur une densité de probabilité de date
}
