% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R
\name{RenDate.plot}
\alias{RenDate.plot}
\title{Trace des courbe de densité}
\usage{
RenDate.plot(
  x,
  withPositions = FALSE,
  pause = FALSE,
  dateHeight = 30,
  borderCol = NULL,
  fillCols = rep("gray", length(x)),
  withHDR = TRUE,
  hdrCol = "darkgray",
  ...
)
}
\arguments{
\item{withHDR}{Calcul le hdr (hpd) et remplie la surface correspondant}

\item{dateHeigth}{Fixe la hauteur des densités}
}
\description{
Trace des courbe de densité
}
