% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R
\name{wiggle.gauss}
\alias{wiggle.gauss}
\title{Calcul le wiggle (décalage) d'une densité de class "RenDate"
Fonction utilisée avec produit.RenDate() pour calculer le "Wiggle Matching"
Il s'agit d'un produit de convolution de la datation par une "densité Normale"}
\usage{
wiggle.gauss(x, mean, sd)
}
\arguments{
\item{x}{une densité (datation) de class "RenDate"}

\item{mean}{la valeur moyenne}

\item{sd}{la valeur de l'écart type}
}
\description{
Calcul le wiggle (décalage) d'une densité de class "RenDate"
Fonction utilisée avec produit.RenDate() pour calculer le "Wiggle Matching"
Il s'agit d'un produit de convolution de la datation par une "densité Normale"
}
