% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R
\name{wiggle.uniform}
\alias{wiggle.uniform}
\title{Calcul le wiggle (décalage) d'une densité de class "RenDate"
Fonction utilisée avec produit.RenDate() pour calculer le "Wiggle Matching"
Il s'agit d'un poroduit de convolution de la datation par une "fonction porte"}
\usage{
wiggle.uniform(x, wiggle.min, wiggle.max)
}
\arguments{
\item{x}{une densité (datation) de class "RenDate"}

\item{wiggle.min}{la borne inférieure du décalage}

\item{wiggle.max}{la borne supérieure du décalage}
}
\description{
Calcul le wiggle (décalage) d'une densité de class "RenDate"
Fonction utilisée avec produit.RenDate() pour calculer le "Wiggle Matching"
Il s'agit d'un poroduit de convolution de la datation par une "fonction porte"
}
