% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R, R/RenDate_20250701.R
\name{calibrate}
\alias{calibrate}
\title{calibrate}
\usage{
calibrate(
  mesures,
  std,
  calCurves,
  ids = NULL,
  positions = NULL,
  pathToCalCurves = paste0(getwd(), "/", "data"),
  timeScale = 1
)

calibrate(
  mesures,
  std,
  calCurves,
  ids = NULL,
  positions = NULL,
  pathToCalCurves = paste0(getwd(), "/", "data"),
  timeScale = 1
)
}
\arguments{
\item{mesures}{mesure ou liste des mesures à calibrer}

\item{std}{erreur ou liste des erreurs sur les mesures à calibrer}

\item{timeScale}{pas de la grille de temps}
}
\description{
{Fonction qui permet le calcul de la densité de probabilté d'une date}

{Fonction qui permet le calcul de la densité de probabilté d'une date}
}
