% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R
\name{calibrate_uniform}
\alias{calibrate_uniform}
\title{calibrate_uniform}
\usage{
calibrate_uniform(
  measure.min,
  measure.max,
  calCurves,
  ids = NULL,
  positions = NULL,
  pathToCalCurves = paste0(getwd(), "/", "data"),
  timeScale = 1
)
}
\arguments{
\item{measure.min}{valeur basse de mesure ou liste des mesures à calibrer}

\item{measure.max}{valeur haute de mesure ou liste des valeurs à calibrer}

\item{timeScale}{pas de la grille de temps}
}
\description{
{Fonction qui permet le calcul de la densité de probabilté d'une date, à partir d'une mesure avec une densité de probabilité uniforme}
}
