% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R
\name{courbe_enveloppe}
\alias{courbe_enveloppe}
\title{Trace une courbe avec son enveloppe d erreur à 1 sigma et deux sigma}
\usage{
courbe_enveloppe(
  t,
  mean,
  std,
  col.env = "forestgreen",
  xlim = NULL,
  ylim = NULL,
  new = TRUE,
  ...
)
}
\description{
Trace une courbe avec son enveloppe d erreur à 1 sigma et deux sigma
}
