% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R
\name{createCalCurve}
\alias{createCalCurve}
\title{createCalCurve}
\usage{
createCalCurve(
  name,
  cal_ages,
  uncal_ages,
  one_sigma = rep(0, length(cal_ages)),
  pathToCalCurves = paste0(getwd(), "/", "data")
)
}
\description{
{Stocke une courbe dans le répertoire / data
Cette fonction est utilisée avec la fonction calibrate.
Inspiré du package Bchron (https://cran.r-project.org/web/packages/Bchron/index.html) développé par Andrew Parnell <Andrew.Parnell at mu.ie>, 
( Haslett J, Parnell AC (2008). “A simple monotone process with application to radiocarbon-dated depth chronologies.” 
Journal of the Royal Statistical Society: Series C (Applied Statistics), 57(4), 399–418.
 https://onlinelibrary.wiley.com/doi/10.1111/j.1467-9876.2008.00623.x/full. )
est modifié pour pouvoir l'utiliser avec des mesures magnétiques, ou d'autres sortes de mesures nécessitants des valeurs décimales}
}
\seealso{
\cite{\code{calibrate} }
}
