% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R
\name{date_uniform}
\alias{date_uniform}
\title{date_uniform}
\usage{
date_uniform(
  gate.min = -0,
  gate.max = 100,
  time.grid.min = -1000,
  time.grid.max = 2000,
  time.grid.scale = 1,
  ids = NULL,
  position = NULL
)
}
\arguments{
\item{gate.min}{début de la porte}

\item{gate.max}{fin de la porte}

\item{ids}{nom ou identifiant}

\item{timeGrid.min}{valeur minimale de la grille}

\item{timeGrid.max}{valeur minimale de la grille}

\item{timeScale}{pas de la grille de temps}
}
\description{
{Fonction qui permet le calcul de la densité de probabilté de date uniforme - fonction porte}
}
