% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R
\name{F14CtoCRA}
\alias{F14CtoCRA}
\title{F14C to CRA}
\usage{
F14CtoCRA(F14C)
}
\arguments{
\item{F14C}{Fraction moderne}
}
\value{
Âge radiocarbone conventionnel (CRA)
}
\description{
Conversion de la fraction moderne (F14C) en CRA (Conventional Radiocarbon Age)
}
