% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R
\name{hpd}
\alias{hpd}
\title{Calcul le hpd (hdr) sur une densité de probabilité de date}
\usage{
hpd(date, prob = 0.95)
}
\arguments{
\item{date}{densité produite par la fonction calibrate, génértant un objet de class "RenDate"}

\item{prob}{requested surface value [0, 1]}
}
\description{
Calcul le hpd (hdr) sur une densité de probabilité de date
}
