% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R, R/RenDate_20250701.R
\name{lines.RenDate}
\alias{lines.RenDate}
\title{lines}
\usage{
\method{lines}{RenDate}(
  x,
  withPositions = FALSE,
  pause = FALSE,
  dateHeight = 30,
  normalize = FALSE,
  borderCol = NULL,
  fillCols = rep("gray", length(x)),
  withHDR = TRUE,
  hdrCol = "darkgray",
  ...
)

\method{lines}{RenDate}(
  x,
  withPositions = FALSE,
  pause = FALSE,
  dateHeight = 30,
  normalize = FALSE,
  borderCol = NULL,
  fillCols = rep("gray", length(x)),
  withHDR = TRUE,
  hdrCol = "darkgray",
  ...
)
}
\arguments{
\item{withHDR}{Calcul le hdr (hdp) et remplie la surface correspondant}
}
\description{
{Trace des courbes de densité avec leurs enveloppes d erreur.}

{Trace des courbes de densité avec leurs enveloppes d erreur.}
}
