% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R
\name{plot.RenDate}
\alias{plot.RenDate}
\title{Trace des courbe de densité}
\usage{
\method{plot}{RenDate}(
  x,
  withPositions = FALSE,
  pause = FALSE,
  dateHeight = 30,
  normalize = FALSE,
  borderCol = NULL,
  fillCols = rep("gray", length(x)),
  withHDR = TRUE,
  hdrCol = "darkgray",
  ...
)
}
\arguments{
\item{normalize}{force le maximum à la valeur 1}

\item{withHDR}{Calcul le hdr (hpd) et remplie la surface correspondant}

\item{dateHeigth}{Fixe la hauteur des densités, quand withPositions est TRUE}
}
\description{
Trace des courbe de densité
}
