% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R, R/RenDate_20250701.R
\name{produit}
\alias{produit}
\title{produit}
\usage{
produit(date1, date2, timeScale = 1)

produit(date1, date2, timeScale = 1)
}
\arguments{
\item{date1}{Première densité de classe "RenDate"}

\item{date2}{Deuxième densité de classe "RenDate"}

\item{timeScale}{permet de modifier la grille de temps}
}
\value{
Une nouvelle densité de classe "RenDate" représentant la combinaison des deux densités en entrée
}
\description{
Calcule la combinaison au sens produit de deux densités de classe "RenDate".

{Calcul la combinaison au sens produit de deux densités de class "RenDate"}
}
