% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R, R/RenDate_20250701.R
\name{produit}
\alias{produit}
\title{produit}
\usage{
produit(date1, date2, timeScale = 1)

produit(date1, date2, timeScale = 1)
}
\arguments{
\item{timeScale}{permet de modifier la grille de temps}
}
\description{
{Calcul la combinaison au sens produit de deux densités de class "RenDate"}

{Calcul la combinaison au sens produit de deux densités de class "RenDate"}
}
