% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenDate.R
\name{trace_to_date}
\alias{trace_to_date}
\title{trace_to_date}
\usage{
trace_to_date(
  trace,
  bw = "nrd",
  adjust = 1,
  from,
  to,
  gridLength = 1024,
  kernel = c("gaussian"),
  ids = NULL,
  position = NULL
)
}
\arguments{
\item{trace}{un vecteur history plot provenant de ChronoModel}

\item{bw}{bandwidth}
}
\value{
une date par lissage avec un noyaux, gaussien par defaut
}
\description{
Conversion d'une trace (un vecteur de valeurs) en densité datation pour utilisation avec les autres fonctions du package
}
